コントローラーの修正

+++ app/controllers/items_controller.rb
@@ -37,6 +37,9 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+    unless user_signed_in? && current_user.id == @item.user.id
+      redirect_to root_path
+    end
     if @item.destroy
       redirect_to root_path
     end
